requirements.txt
python_speech_features
resampy
scipy
gdown
tqdm
tensorflow-gan